requirements.txt
# 基础依赖
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 模型与训练
transformers>=4.36.0
accelerate>=0.25.0
deepspeed>=0.12.0
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0

# 多模态
clip @ git+https://github.com/openai/CLIP.git
timm>=0.9.0
einops>=0.7.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
sentencepiece>=0.1.99
tokenizers>=0.15.0

# 数据收集
itchat-uos>=1.5.0.dev0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html2text>=2020.1.16

# 评估与推理
scikit-learn>=1.3.0
nltk>=3.8.0
rouge>=1.0.1
jieba>=0.42.1
sacrebleu>=2.3.0

# 部署服务
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0
streamlit>=1.28.0

# 工具库
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
wandb>=0.16.0
tensorboard>=2.15.0

# 开发工具
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
